requirements.txt
openai>=1.12.0
elevenlabs>=0.2.26
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1 